+++ src/test/java/com/dynatrace/openkit/protocol/BeaconTest.java
@@ -3234,8 +3234,8 @@ public class BeaconTest {
         String encodedUrl = URLEncoder.encode(rawUrl, "UTF-8");
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn(rawUrl);
-        when(webRequestTracer.getBytesSent()).thenReturn(13);
-        when(webRequestTracer.getBytesReceived()).thenReturn(14);
+        when(webRequestTracer.getBytesSent()).thenReturn(13L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(14L);
         when(webRequestTracer.getResponseCode()).thenReturn(15);
 
         final Beacon target = createBeacon().build();
@@ -3316,8 +3316,8 @@ public class BeaconTest {
         String encodedUrl = URLEncoder.encode(rawUrl, "UTF-8");
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn(rawUrl);
-        when(webRequestTracer.getBytesSent()).thenReturn(0);
-        when(webRequestTracer.getBytesReceived()).thenReturn(14);
+        when(webRequestTracer.getBytesSent()).thenReturn(0L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(14L);
         when(webRequestTracer.getResponseCode()).thenReturn(15);
 
         final Beacon target = createBeacon().build();
@@ -3353,8 +3353,8 @@ public class BeaconTest {
         String encodedUrl = URLEncoder.encode(rawUrl, "UTF-8");
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn(rawUrl);
-        when(webRequestTracer.getBytesSent()).thenReturn(-1);
-        when(webRequestTracer.getBytesReceived()).thenReturn(14);
+        when(webRequestTracer.getBytesSent()).thenReturn(-1L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(14L);
         when(webRequestTracer.getResponseCode()).thenReturn(15);
 
         final Beacon target = createBeacon().build();
@@ -3389,8 +3389,8 @@ public class BeaconTest {
         String encodedUrl = URLEncoder.encode(rawUrl, "UTF-8");
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn(rawUrl);
-        when(webRequestTracer.getBytesSent()).thenReturn(13);
-        when(webRequestTracer.getBytesReceived()).thenReturn(0);
+        when(webRequestTracer.getBytesSent()).thenReturn(13L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(0L);
         when(webRequestTracer.getResponseCode()).thenReturn(15);
 
         final Beacon target = createBeacon().build();
@@ -3426,8 +3426,8 @@ public class BeaconTest {
         String encodedUrl = URLEncoder.encode(rawUrl, "UTF-8");
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn(rawUrl);
-        when(webRequestTracer.getBytesSent()).thenReturn(13);
-        when(webRequestTracer.getBytesReceived()).thenReturn(-1);
+        when(webRequestTracer.getBytesSent()).thenReturn(13L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(-1L);
         when(webRequestTracer.getResponseCode()).thenReturn(15);
 
         final Beacon target = createBeacon().build();
@@ -3462,8 +3462,8 @@ public class BeaconTest {
         String encodedUrl = URLEncoder.encode(rawUrl, "UTF-8");
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn(rawUrl);
-        when(webRequestTracer.getBytesSent()).thenReturn(13);
-        when(webRequestTracer.getBytesReceived()).thenReturn(14);
+        when(webRequestTracer.getBytesSent()).thenReturn(13L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(14L);
         when(webRequestTracer.getResponseCode()).thenReturn(0);
 
         final Beacon target = createBeacon().build();
@@ -3499,8 +3499,8 @@ public class BeaconTest {
         String encodedUrl = URLEncoder.encode(rawUrl, "UTF-8");
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn(rawUrl);
-        when(webRequestTracer.getBytesSent()).thenReturn(13);
-        when(webRequestTracer.getBytesReceived()).thenReturn(14);
+        when(webRequestTracer.getBytesSent()).thenReturn(13L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(14L);
         when(webRequestTracer.getResponseCode()).thenReturn(-1);
 
         final Beacon target = createBeacon().build();
@@ -3533,8 +3533,8 @@ public class BeaconTest {
         // given
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn("https://www.google.com");
-        when(webRequestTracer.getBytesSent()).thenReturn(13);
-        when(webRequestTracer.getBytesReceived()).thenReturn(14);
+        when(webRequestTracer.getBytesSent()).thenReturn(13L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(14L);
         when(webRequestTracer.getResponseCode()).thenReturn(15);
 
         when(mockServerConfiguration.isSendingDataAllowed()).thenReturn(false);
@@ -3553,8 +3553,8 @@ public class BeaconTest {
         // given
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn("https://www.google.com");
-        when(webRequestTracer.getBytesSent()).thenReturn(13);
-        when(webRequestTracer.getBytesReceived()).thenReturn(14);
+        when(webRequestTracer.getBytesSent()).thenReturn(13L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(14L);
         when(webRequestTracer.getResponseCode()).thenReturn(15);
 
         when(mockPrivacyConfiguration.isWebRequestTracingAllowed()).thenReturn(false);
@@ -3574,8 +3574,8 @@ public class BeaconTest {
         // given
         WebRequestTracerBaseImpl webRequestTracer = mock(WebRequestTracerBaseImpl.class);
         when(webRequestTracer.getURL()).thenReturn("https://www.google.com");
-        when(webRequestTracer.getBytesSent()).thenReturn(13);
-        when(webRequestTracer.getBytesReceived()).thenReturn(14);
+        when(webRequestTracer.getBytesSent()).thenReturn(13L);
+        when(webRequestTracer.getBytesReceived()).thenReturn(14L);
         when(webRequestTracer.getResponseCode()).thenReturn(15);
 
         int trafficControlPercentage = 50;